zarchives/apps/search/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
arrow==1.2.1
    # via -r requirements.in
attrs==20.2.0
    # via pytest
azure-core==1.8.1
    # via
    #   azure-cosmos
    #   azure-storage-blob
azure-cosmos==4.1.0
    # via -r requirements.in
azure-storage-blob==12.5.0
    # via -r requirements.in
certifi==2020.6.20
    # via
    #   msrest
    #   requests
cffi==1.14.3
    # via cryptography
chardet==3.0.4
    # via requests
cryptography==3.1.1
    # via azure-storage-blob
docopt==0.6.2
    # via -r requirements.in
idna==2.10
    # via requests
iniconfig==1.0.1
    # via pytest
isodate==0.6.0
    # via msrest
more-itertools==8.5.0
    # via pytest
msrest==0.6.19
    # via azure-storage-blob
oauthlib==3.1.0
    # via requests-oauthlib
packaging==20.4
    # via pytest
pluggy==0.13.1
    # via pytest
py==1.9.0
    # via pytest
pycparser==2.20
    # via cffi
pyparsing==2.4.7
    # via packaging
pytest==6.0.2
    # via -r requirements.in
python-dateutil==2.8.2
    # via arrow
requests==2.24.0
    # via
    #   -r requirements.in
    #   azure-core
    #   msrest
    #   requests-oauthlib
requests-oauthlib==1.3.0
    # via msrest
six==1.15.0
    # via
    #   azure-core
    #   azure-cosmos
    #   cryptography
    #   isodate
    #   packaging
    #   python-dateutil
toml==0.10.1
    # via pytest
urllib3==1.25.10
    # via requests
